web: SettingsLabgrid: break out ConfigEditor and SystemdServices
These components are generally useful (e.g. in the setup wizard, as you will
see soon). Break them out into separate files.
